include kernalization before branching
