+++ vertex_cover_solver.py
@@ -1,11 +1,60 @@
 import sys
 
-import reading
-
 g = {}
 max_degree = 0
 degree_list = []
 nb_vertices = 0
+
+def add_vertex(vertex):
+    """
+    INPUT: g is dict with each value list of length 3 (boolean, int, list), vertex is str
+    add_vertex adds a new vertex with default values
+    OUTPUT: dict with each value list of 3 (boolean, int, list)
+    """
+    g[vertex] = [False, 0, []]
+
+
+def add_edge(edge):
+    """
+    INPUT: g is dict with each value list of length 3 (boolean, int, list), edge is list of length 2
+    add_vertex adds a new edge to the graph and returns this graph
+    OUTPUT: dict with each value list of 3 (boolean, int, list)
+    """
+    global max_degree
+    for vertex in edge:
+        if not vertex in g.keys():
+            add_vertex(vertex)
+        g[vertex][1] += 1
+        # If current degree is greater than maximum degree, update:
+        if g[vertex][1] > max_degree:
+            max_degree += 1
+    g[edge[0]][2].append(edge[1])
+    g[edge[1]][2].append(edge[0])
+
+
+def get_data():
+    """
+    INPUT: None
+    get_data reads standard input and creates the given graph
+    OUTPUT: None
+    """
+    global degree_list
+    global nb_vertices
+    # Get standard input:
+    input_data = sys.stdin
+    for line in input_data:
+        if not line[0] == '#':
+            # Get current edge and add it to the graph:
+            current_edge = line.split()
+            add_edge(current_edge)
+    # Initializing degree_list:
+    degree_list = [[] for i in range(max_degree + 1)]
+    for vertex in g:
+        degree = g[vertex][1]
+        # Append vertex to the list located at its degree in degree_list:
+        degree_list[degree].append(vertex)
+    nb_vertices = len(g)
+
 
 def print_result(vertices):
     """
@@ -244,23 +293,21 @@ def vc():
     OUTPUT:None, prints directly in the console
     """
     vc_branch.counter = 0
-    # Get neighbors of vertices with degree one (if two are adjacent to each other, only one of them):
-    degree_one_neighbors = get_degree_one_neighbors()
-    # 'Delete' the neighbors from the graph:
-    del_vert(degree_one_neighbors)
-    if is_edgeless():
-        S = degree_one_neighbors
+    if is_edgeless(): S = []
     else:
-        kmin = bound()
-        for k in range(kmin,len(g)):
-            S = vc_branch(k)
-            if S is not None:
-                S += degree_one_neighbors
-                break
+        S_kern, _, _ = kernalization(len(g) - 1)
+        if is_edgeless(): S = S_kern
+        else:
+            kmin = bound()
+            for k in range(kmin, len(g)):
+                S = vc_branch(k)
+                if S is not None:
+                    S += S_kern
+                    break
     print_result(S)
     print("#recursive steps: %s" % vc_branch.counter)
     return None
 
 
-reading.get_data()
+get_data()
 vc()
